+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [4.0.0] - 2024-05-10
+
+### Added
+
+Add a metadata store to hold the project trigram and the runtime environment.
+
+### How to migrate
+
+Add the `trigram` variable containing the project trigram to the module.
+
+
 ## [3.0.0] - 2024-04-08
 
 ### Breaking change